change architecture in order to fit our architectural plans with schlucht
