+++ .gitignore
@@ -51,4 +51,5 @@ Temporary Items
 
 config.yml
 
-*.dia.+
+*.dia
+*.png
