+++ destiny/main/bdd/models/__init__.py
@@ -1,0 +1,8 @@
+from destiny.main.bdd.models.players import Players
+from destiny.main.bdd.models.participant import Participant
+from destiny.main.bdd.models.assistevent import AssistEvent
+from destiny.main.bdd.models.itemevent import ItemEvent
+from destiny.main.bdd.models.stats import Stats
+from destiny.main.bdd.models.victimEvent import VictimEvent
+from destiny.main.bdd.models.killevent import KillEvent
+from destiny.main.bdd.models.matches import Matches
