+++ destiny/main/bdd/models/assistevent.py
@@ -1,14 +1,14 @@
 from sqlalchemy import Column
 from sqlalchemy.sql.sqltypes import BigInteger, CHAR, Integer
 
-from models.base import Base, rep_model
+from destiny.main.bdd.base import Base, rep_model
 
 
-class KillEvent(Base):
-    __tablename__ = 'killEvent'
+class AssistEvent(Base):
+    __tablename__ = 'assistEvent'
 
     gameId = Column(BigInteger, nullable=False, primary_key=True)
-    killer = Column(CHAR(50), primary_key=True)
+    assist = Column(CHAR(50), primary_key=True)
     victim = Column(CHAR(50), primary_key=True)
     timestamp = Column(BigInteger, primary_key=True)
     x = Column(Integer)
@@ -18,5 +18,5 @@ class KillEvent(Base):
         return rep_model(self)
 
 if __name__ == '__main__':
-    a = KillEvent()
+    a = AssistEvent()
     print(a)
