+++ destiny/main/bdd/models/killevent.py
@@ -1,14 +1,14 @@
 from sqlalchemy import Column
 from sqlalchemy.sql.sqltypes import BigInteger, CHAR, Integer
 
-from models.base import Base, rep_model
+from destiny.main.bdd.base import Base, rep_model
 
 
-class AssistEvent(Base):
-    __tablename__ = 'assistEvent'
+class KillEvent(Base):
+    __tablename__ = 'killEvent'
 
     gameId = Column(BigInteger, nullable=False, primary_key=True)
-    assist = Column(CHAR(50), primary_key=True)
+    killer = Column(CHAR(50), primary_key=True)
     victim = Column(CHAR(50), primary_key=True)
     timestamp = Column(BigInteger, primary_key=True)
     x = Column(Integer)
@@ -18,5 +18,5 @@ class AssistEvent(Base):
         return rep_model(self)
 
 if __name__ == '__main__':
-    a = AssistEvent()
+    a = KillEvent()
     print(a)
