--- models/__init__.py
@@ -1,8 +1,0 @@
-from models.players import Players
-from models.participant import Participant
-from models.assistevent import AssistEvent
-from models.itemevent import ItemEvent
-from models.stats import Stats
-from models.victimEvent import VictimEvent
-from models.killevent import KillEvent
-from models.matches import Matches
